Refactored to have a separate implementation for ValidationRuleEngine
