+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/EtorDomainRegistration.java
@@ -33,9 +33,7 @@ import gov.hhs.cdc.trustedintermediary.etor.results.ResultConverter;
 import gov.hhs.cdc.trustedintermediary.etor.results.ResultResponse;
 import gov.hhs.cdc.trustedintermediary.etor.results.ResultSender;
 import gov.hhs.cdc.trustedintermediary.etor.results.SendResultUseCase;
-import gov.hhs.cdc.trustedintermediary.etor.ruleengine.RuleEngine;
 import gov.hhs.cdc.trustedintermediary.etor.ruleengine.RuleLoader;
-import gov.hhs.cdc.trustedintermediary.etor.ruleengine.ValidationRule;
 import gov.hhs.cdc.trustedintermediary.etor.ruleengine.ValidationRuleEngine;
 import gov.hhs.cdc.trustedintermediary.external.database.DatabaseMessageLinkStorage;
 import gov.hhs.cdc.trustedintermediary.external.database.DatabasePartnerMetadataStorage;
@@ -137,7 +135,7 @@ public class EtorDomainRegistration implements DomainConnector {
         ApplicationContext.register(RuleLoader.class, RuleLoader.getInstance());
         ApplicationContext.register(
                 ValidationRuleEngine.class,
-                RuleEngine.getInstance("validation_definitions.json", ValidationRule.class));
+                ValidationRuleEngine.getInstance("validation_definitions.json"));
 
         ApplicationContext.register(SendMessageHelper.class, SendMessageHelper.getInstance());
