--- etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/ruleengine/IRuleEngine.java
@@ -1,10 +1,0 @@
-package gov.hhs.cdc.trustedintermediary.etor.ruleengine;
-
-public interface IRuleEngine {
-
-    void unloadRules();
-
-    void ensureRulesLoaded();
-
-    void runRules(FhirResource<?> resource);
-}
